closes #24: - loop that writes to DMA still waiting for too many samples. previously I would keep the acq. time fixed and send all the measurements to the host which would then do the averaging. Now I just send the averaged value to the host, thus I send less samples.

+++ src/labview_fpga_lib/galvo_scan/galvo_scan.py
@@ -48,7 +48,7 @@ getter_functions = {
     "read_DMA_elem_to_write": c_uint32,
     "read_DMATimeOut": c_bool,
     "read_elements_written_to_dma": c_int32,
-    "read_detector_signal": c_int16,
+    "read_detector_signal": c_int32,
     "read_ix": c_int32,
     "read_iy": c_int32,
     "read_Nx": c_int32,
@@ -92,14 +92,14 @@ def stop_FIFO(session, status):
 
 
 # read FIFO
-_libfpga.read_FIFO.argtypes = [POINTER(c_int16), c_uint32,
+_libfpga.read_FIFO.argtypes = [POINTER(c_int32), c_uint32,
                                POINTER(c_uint32), POINTER(c_int32),
                                POINTER(c_uint32)]
 
 _libfpga.read_FIFO.restype = None
 
 def read_FIFO(size, session, status):
-    Signal = (c_int16*size)()
+    Signal = (c_int32*size)()
     elements_remaining = c_uint32()
 
     _libfpga.read_FIFO(Signal, size, byref(session), byref(status), byref(elements_remaining))
